requirements.txt
ffmpeg-python==0.2.0
keras==2.7.0
librosa==0.9.2
numpy==1.19.5
streamlit==1.11.1
tensorflow==2.7.0
